Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Want to throw in a +1 for this. It's very hard to trigger on an unmodified Launcher that doesn't auto-orientate, but it's still a nasty bug and this definitely fixes it.

Patch-set: 2
Reviewer: Steve Kondik <1002408@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1