Update patch set 2
Patch Set 2: No score; No score

Could we take the change in and ignore/skip/override it when we import it into Google's internal tree? It'd fix the issue for now, and it'd give Christopher credit by still having his change in the main history.

(I'm OK special-casing this change in the import/merge process to make that happen)

Patch-set: 2