Update patch set 2
Actually I was a bit hasty in my assessment. Never do code reviews so early after returning from vacations :) There is a race condition in this code, yes, and it was fixed in 2.0. However, this patch is not correct: the boolean that are protected by a synchronized block in this patch are volatile. There is no reason to synchronize before reading/writing them.

Patch-set: 2