Update patch set 2
The code flow, as I understand it, is that in the Loader thread, mStopped was being checked, then the other thread got its timeslice, where mStopped was being set *and* m{Applications,DesktopItems}Loaded was being set to false, then control returned to the Loader thread, where m{Applications,DesktopItems}Loaded was set to true.

I am not very familiar with the code, so I don't know if this was the race condition causing the problems (or even if it could cause any), but I do know that it was a race condition, and it appeared to be triggered looking at my log output from the case that I could reproduce.

Patch-set: 2