Update patch set 2
Patch Set 2: Verified; I would prefer that you didn't submit this

I'm not quite convinced by the change.

If I'm reading correctly, in each of the pairs of synchronized blocks, one of the two blocks is a single plain access to a volatile variable. From the point of view of strict synchronization, I don't believe that this change makes a difference.

However, explicitly synchronization is heavier than volatile access and is likely to change the timing of the code (possibly by interacting with the scheduler), so it's very possible that this change makes the symptom disappear as a side effect.

Patch-set: 2
Label: Code-Review=-1
Label: Verified=+1