Update patch set 2
The race it fixes (which, again, may have no bearing on operation) does not occur due to mStopped (which is volatile), but rather due to the way mStopped is used to interact with m{Applications,DesktopItems}Loaded.

mStopped access is atomic, but checking mStopped then setting m{Applications,DesktopItems}Loaded based on that value is not.

Setting mStopped then m{Applications,DesktopItems}Loaded is also not atomic.

A race can happen (the one I described in an earlier comment) when the two non-atomic operations happen near-simultaneously.

The synchronizing blocks make the former operation (check, maybe set) atomic - The latter (set, set) does not need to be atomic if the former is.

Patch-set: 2